+++ private-components/swingui/src/test/java/org/limewire/ui/swing/search/model/MockVisualSearchResult.java
@@ -23,7 +23,7 @@ public class MockVisualSearchResult implements VisualSearchResult {
     private VisualSearchResult similarityParent;
     private boolean spam;
     private HashMap<FilePropertyKey, Object> properties = new HashMap<FilePropertyKey, Object>();
-    private double relevance = 0;
+    private int relevance = 0;
     
     public MockVisualSearchResult(String name) {
         this.name = name;
@@ -217,11 +217,11 @@ public class MockVisualSearchResult implements VisualSearchResult {
     }
 
     @Override
-    public double getRelevance() {
+    public int getRelevance() {
         return relevance;
     }
     
-    public void setRelevance(double relevance) {
+    public void setRelevance(int relevance) {
         this.relevance = relevance;
     }
